Create/update user,item,account Controllers, Services, Models

+++ src/main/java/com/jin/game/repository/UserRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.math.BigInteger;
 import java.util.List;
 
-public interface UserRepository extends JpaRepository<User, BigInteger> {
+public interface UserRepository extends JpaRepository<User, Long> {
     List<User> findByUserName(String username);
 }
